Integrate Authentication Middleware
