+++ backend/src/index.ts
@@ -1,12 +1,16 @@
+import { Hono } from 'hono'
 import { logger } from 'hono/logger'
 
 import { errorMiddleware } from 'wildebeest/backend/src/middleware'
+import { HonoEnv } from 'wildebeest/backend/src/types'
 import { createApp } from 'wildebeest/backend/src/utils'
 
-const app = createApp()
+const base = new Hono<HonoEnv>()
 
-app.use('*', logger())
-app.use('*', errorMiddleware())
+base.use('*', logger())
+base.use('*', errorMiddleware())
+
+const app = createApp({ app: base })
 
 app.showRoutes()
