+++ backend/src/middleware/index.ts
@@ -1,3 +1,2 @@
 export { corsMiddleware } from './cors'
 export { errorMiddleware } from './error'
-export { publicMiddleware, privateMiddleware } from './auth'
